# terraform/environments/dev/api-spec.yaml
swagger: "2.0"
info:
  title: "Pixel Canvas API"
  description: "API Gateway for serverless pixel canvas application"
  version: "2.0.0"
schemes:
  - "https"
produces:
  - "application/json"
x-google-endpoints:
  - name: "pixel-canvas-gateway-86fcxr1p.ew.gateway.dev"
    allowCors: true
paths:
  /discord/webhook:
    post:
      summary: "Discord webhook endpoint"
      operationId: "discordWebhook"
      x-google-backend:
        address: "${discord_proxy_url}"
        protocol: "h2"
      responses:
        200:
          description: "Success"
        401:
          description: "Unauthorized"
        500:
          description: "Internal server error"
  /auth/login:
    get:
      summary: "Initiate Discord OAuth2 login"
      operationId: "authLogin"
      x-google-backend:
        address: "${auth_handler_url}/auth/login"
        protocol: "h2"
      responses:
        302:
          description: "Redirect to Discord OAuth2"
        500:
          description: "Internal server error"
  /auth/callback:
    get:
      summary: "OAuth2 callback endpoint"
      operationId: "authCallback"
      x-google-backend:
        address: "${auth_handler_url}/auth/callback"
        protocol: "h2"
      parameters:
        - name: code
          in: query
          type: string
          required: true
        - name: state
          in: query
          type: string
          required: false
      responses:
        200:
          description: "Authentication successful, returns JWT"
        400:
          description: "Missing authorization code"
        500:
          description: "Authentication failed"
  /auth/me:
    get:
      summary: "Get current user info"
      operationId: "authMe"
      x-google-backend:
        address: "${auth_handler_url}/auth/me"
        protocol: "h2"
      responses:
        200:
          description: "User info"
        401:
          description: "Unauthorized"
        404:
          description: "User not found"
    options:
      summary: "CORS preflight for /auth/me"
      operationId: "authMeCors"
      x-google-backend:
        address: "${auth_handler_url}/auth/me"
        protocol: "h2"
      responses:
        204:
          description: "CORS preflight response"
  /api/pixels:
    get:
      summary: "Get all pixels"
      operationId: "getPixels"
      x-google-backend:
        address: "${web_proxy_url}/api/pixels"
        protocol: "h2"
      responses:
        200:
          description: "List of pixels"
        500:
          description: "Internal server error"
    post:
      summary: "Place a pixel"
      operationId: "placePixel"
      x-google-backend:
        address: "${web_proxy_url}/api/pixels"
        protocol: "h2"
      parameters:
        - name: body
          in: body
          required: true
          schema:
            type: object
            properties:
              x:
                type: integer
              y:
                type: integer
              color:
                type: string
      responses:
        202:
          description: "Pixel placement accepted"
        401:
          description: "Unauthorized"
        400:
          description: "Invalid pixel data"
    options:
      summary: "CORS preflight for /api/pixels"
      operationId: "pixelsCors"
      security: []
      x-google-backend:
        address: "${web_proxy_url}/api/pixels"
        protocol: "h2"
        disable_auth: true
      responses:
        204:
          description: "CORS preflight response"
  /api/canvas:
    get:
      summary: "Get canvas state"
      operationId: "getCanvas"
      x-google-backend:
        address: "${web_proxy_url}/api/canvas"
        protocol: "h2"
      responses:
        200:
          description: "Canvas state"
        500:
          description: "Internal server error"
    options:
      summary: "CORS preflight for /api/canvas"
      operationId: "canvasCors"
      security: []
      x-google-backend:
        address: "${web_proxy_url}/api/canvas"
        protocol: "h2"
        disable_auth: true
      responses:
        204:
          description: "CORS preflight response"